refactor: extract read_status_engine to internal/readstatus (P5)
Moves RecomputeUserBookState + SetManualStatus from internal/server/
into a new internal/readstatus/ package. Unblocks the iTunes service
extraction — internal/server/itunes_position_sync.go was reaching into
server-package code for these helpers, which would have been a circular
dep once the file moved to internal/itunes/service/ (server already
imports itunes/service).

Pure file move — signatures unchanged. Both functions already accepted
narrow interface types (BookFileStore + UserPositionStore) from the
ISP sweep, so no new Deps wiring needed. Tests moved with the file.

Callers updated: 2 files (internal/server/itunes_position_sync.go,
internal/server/reading_handlers.go) — 5 call sites total now
qualified with readstatus. prefix.

Pre-work for the iTunes extraction Phase 2 M1 (spec v2 §5.1).
